% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longwing-package.R
\docType{package}
\name{longwing-package}
\alias{longwing}
\alias{longwing-package}
\title{longwing: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wzbillings/longwing}
  \item \url{https://wzbillings.github.io/longwing/}
  \item Report bugs at \url{https://github.com/wzbillings/longwing/issues}
}

}
\keyword{internal}
